pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "dependency-injector[pydantic2]>=4.48.2",
    "fastapi[standard]>=0.118.2",
    "langchain>=0.3.27",
    "langchain-postgres>=0.0.15",
    "pwdlib[argon2]>=0.2.1",
    "pydantic>=2.11.10",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "sqlalchemy[asyncio]>=2.0.43",
    "uvicorn[standard]>=0.37.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
]
